chore: refactoring frontend for lint
